+++ ali/common/src/main/java/com/yanny/ali/network/AbstractServer.java
@@ -387,7 +387,7 @@ public abstract class AbstractServer {
 
                 if (itemListingMap != null && itemListingMap.int2ObjectEntrySet().stream().anyMatch((e) -> e.getValue().length > 0)) {
                     try {
-                        nodes.put(location, serverRegistry.parseTrade(itemListingMap));
+                        nodes.put(location, serverRegistry.parseTrade(itemListingMap, false));
                         tradeItems.put(location, ItemCollectorUtils.collectTradeItems(serverRegistry, itemListingMap));
                     } catch (Throwable e) {
                         LOGGER.warn("Failed to parse trade for villager {} with error {}", location, e.getMessage(), e);
@@ -410,9 +410,9 @@ public abstract class AbstractServer {
                 Int2ObjectMap<VillagerTrades.ItemListing[]> trades = new Int2ObjectOpenHashMap<>();
 
                 VillagerTrades.EXPERIMENTAL_WANDERING_TRADER_TRADES.forEach((pair) -> trades.put((int) pair.getValue(), pair.getKey()));
-                return serverRegistry.parseTrade(trades);
+                return serverRegistry.parseTrade(trades, true);
             } else {
-                return serverRegistry.parseTrade(VillagerTrades.WANDERING_TRADER_TRADES);
+                return serverRegistry.parseTrade(VillagerTrades.WANDERING_TRADER_TRADES, true);
             }
         } catch (Throwable e) {
             LOGGER.warn("Failed to parse wandering trader with error {}", e.getMessage(), e);
@@ -529,7 +529,7 @@ public abstract class AbstractServer {
 
             // write dummy data
             buf.writerIndex(wtStart);
-            new TradeNode(utils, new Int2ObjectOpenHashMap<>()).encode(utils, buf);
+            new TradeNode(utils, new Int2ObjectOpenHashMap<>(), true).encode(utils, buf);
             buf.writeCollection(List.of(), (b, i) -> {});
             buf.writeCollection(List.of(), (b, i) -> {});
         } finally {